Refactor logging to allow optional date prefix
When the logging output is being sent to the journald, we want to avoid
having the datetime prefix added by the n3n logging - the log lines will
end up in the journal with both a n3n datestamp and a journal datestamp.

While adding this, refactor the function to remove duplication and
various overheads.  Reduces the stack usage and the number of times the
strings are scanned.