Hacky conversion of transop id to use new parser.
Eventually, I want to have each available transop register itself with a
transop manager interface (similar to how the n3n_config_register_section
is used to register available config options), however that is not ready
yet.

In a config file, I definitely want to use a descriptive string to
configure the cipher used, so these numbers will be going away.

The current commandline option will probably be deprecated, too, which
would mean that nothing external is using a raw ID number for any cipher
settings.

For now, the less of evils seems to be to quick check against the bounds
of the transform id.

To ensure we have a known memory storage size for the config parser,
this also changes the transop_id to be uint8_t.