Add compressions to the transform registration.
Note that both encryption and compression are handled internally with
transop structures, but they have a different ID namespace with
overlapping ID numbers - which is unfortunate, since we cannot simply
use the one transform system.

Hopefully, this can get untangled a little better in the future, but
the obvious route would require an incompatible protocol change.