Allow proper config overwriting with environment
To allow a clear order of precedence for the overiding of config
variables, the setting config from the environment is separated out
into its own function.

This allows the commandline args to overide the environment and the
environment to overide the config file and the config file to override
the defaults.