Produce quieter makefile output
Since the number of options being given to each compile is probably only
going to increase, the ability to read and understand the compile output
is only going to become harder.

Address this by outputing the generic CC line once and then just show
the name of each file that is being compiled.

Note that we deliberately do not do this for the apps/ directory as we
want that to stay as simple as possible so that people can easily use it
as a template to write their own app.