Merge the supernode and edge data structures
This is simply the first pass to allow things to work.  More work needs
to be done to properly merge things.  Many of the supernode runtime data
items are actually conf - these should be sorted out during the
migration to the new config parser.

This change will allow the sharing of the management interface, which
will allow the old UDP API to be removed.  It will also simplify the CLI
parsing cleanup, the use of a sessiondir and the config file parser. It
should help with any future merging of the supernode and edge functions
into one daemon, but that is a long way away yet.