The default is now to run in foreground.
Since most people who will be running the processes in the background
will be doing this with some kind of process supervisor - which will be
assisted by the daemon not forking, we can make the default case
simpler.

This also has the helpful side effect of allowing the removal of one
more non-zero-default variable in the *_defaults() function.