Use correct initial version in changelog
Since we wish to use the git describe feature to create the versions,
and that generates versions with dashes, and debian has specific
requirements for versions that have dashes, we need to make the correct
version number show up in the debian changelog.

This is unfortunate, since it also means that we cannot call this
package "native", however consistently using the exact VCS description
is significantly more useful.