Dont use encode functions when not needed
The wire encode_* functions are intended to build the packets that will
be sent on the wire.  These uses of encode_mac() were for structures
that would then be passed to another encode_* function, thus they are
not building wire structs.

Additionally, the encode_mac() degenerates into a string copy, so just
do that.