Ensure the transform tests using stable nonces
Add a function to the random number generator to reseed with a fixed and
known seed - this allows us to run any tests that consume random numbers
with a known and expected result.

Specifically, the transform functions use a prefix of randomness that is
basically serving as a nonce.

This was previously stable because we didnt seed the random number
generator, but adding any new transform functions in the middle would
have perturbed the following results - so we use the stable default
before each transform.

This came to light as the result of changing to use native host ordering
for the random nonce bytes - which disturbed the encrypted output.