Avoid showing known warning compiling speck
The speck code has been generating warnings at compile time for as long
as I can remember.  Whilst it may be a valid warning, it does not appear
to have random uninitialised data in the Y[2] as the unit tests are
producing stable results.

Ideally, this warning would be fixed.  The simple fix of initialising
the variable could slow down the encryption routine.  So - before trying
that - there should be an automated encryption performance test added.

There is also a plan to make more accessible optimised versions of the
crypto routines, which would require refactoring this code, so it may
unearth more warnings or shift the message in a way that helps to
diagnose things.

For the moment, take the simple path of ignoring that warning for this
one file.

```
In file included from /usr/lib/gcc/x86_64-linux-gnu/12/include/xmmintrin.h:1316,
                 from /usr/lib/gcc/x86_64-linux-gnu/12/include/immintrin.h:31,
                 from /home/hamish/r/n3n/include/speck.h:75,
                 from src/speck.c:25:
In function ‘_mm_unpackhi_epi64’,
    inlined from ‘speck_encrypt_xor.isra’ at src/speck.c:587:9:
/usr/lib/gcc/x86_64-linux-gnu/12/include/emmintrin.h:1024:19: warning: ‘Y[2]’ may be used uninitialized [-Wmaybe-uninitialized]
 1024 |   return (__m128i)__builtin_ia32_punpckhqdq128 ((__v2di)__A, (__v2di)__B);
      |                   ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
src/speck.c: In function ‘speck_encrypt_xor.isra’:
src/speck.c:540:16: note: ‘Y[2]’ was declared here
  540 |     u128 X[4], Y[4], Z[4];                                 \
      |                ^
src/speck.c:587:9: note: in expansion of macro ‘Encrypt_Dispatcher’
  587 |         Encrypt_Dispatcher(128);
      |         ^~~~~~~~~~~~~~~~~~
```