Attempt to allow useful install dirs
This plumbs in the normal autoconf settings for the docdir and mandir.
These autoconf tools dont have an appropriate dir location for the
systemd unit files however (the libdir is intended for exec files, not
config files) so there is a new --with-systemddir=DIR option to set
that with some heuristics to set (hopefully) reasonable default dirs.

Addresses #14 and #15