Rename applications to use the project name
By prefixing the two main application binaries ("n3n-edge" and
"n3n-supernode") with the name of the project, it makes it clear that
these generic names are associated with a specific project.

This change has been suggested in the past, but was always avoided as it
was a configuration-breaking change.  Since the new n3n release is
making other breaking changes, now is the right time to perform this
rename.