Use working commitish for openwrt build
When building a PR merge, github will set GITHUB_REF to a value like
"refs/pull/24/merge", which the openwrt mechanism will attempt to
checkout, but that string is not acceptable.

Switch to always using the commit hash as that should always be one that
the openwrt git checkout process can find.