More messing with autotools and dirty builds
Some times the autoreconf will create the required files and sometimes
it will just complain that the files are missing.  If we include the
files as part of the repo to avoid the missing errors, then we will
often have a dirty repository when doing builds.

Walk a middle line, trying to have our cake and not lie about it.