Be more careful with selected file descriptors
Some bug reports that edges using TCP mode are getting occasional
errors:
   *** bit out of range 0 - FD_SETSIZE on fd_set ***: terminated

We were not accounting for the fact that the TCP supernode socket file
descriptor is set to "-1" when it is disconnected and invalid.

Avoid "-1" file handles when adding to or checking signals for fd_set
contents.

This code path will hopefully also get smoothed out when the mainloop
conversion is added in - as it will have a systematic way to add/check
file descriptors.

(Addresses #28)