Fix two bugs with our speck API usage
The speck_init() function will always allocate its own buffer, so dont
allocate one as that leads to memory leaks.

In some configuations, the speck_init will not use the standard malloc()
and this may cause issues when trying to use free() on the returned ctx.

This issue definitely occurs with the Windows builds and lead to the
supernode aborting with "exception code 0xc0000374 (Heap Corruption)"

(Ref #30)