Change peer event type names
The structure of the packet stays the same, but the descriptive strings
change to allow clearer information.  This is technically an API change,
but is intended to be a minor one - there are no known users of these
strings.

TODO: document the event messages better, so that the exact state
transitions are more well known.