Attempt to plumb the TCP status into anything
Nothing in the usual peer_info or n2n_sock_t indicates that a peer is to be
connected via TCP - it is all done by a separate connection table.  This
makes it hard to get status output that shows that a given peer should
be reachable via TCP.

This patch was an attempt to connect some of the dots, sadly it was not
enough.  It is probably useful framework, but more still needs to be
done.

Of most annoyance is that the new TCP test in test_integration_packets.sh
is sending and receiving over TCP, but gets a wire sock structure that
is still flagged as UDP