Fine tune the systemd unit directory.
Assume that any system with systemd has a /usr/lib/systemd dir and
a unit search path within that.  This means we can remove one of the
hacks for the default location of the systemddir.

This can still be overridden by a cmdline arg given to the configure
command.

(Also fix another github changed upload-artifact bug)