Avoid http pipelining more robustly
Check in the write path for finish and cleanup then.
Check in the read select path for an existing transmit and avoid even
starting to read.

This is largely caused because we have no queue of buffers and are often
sharing the same strbuf for ->request and ->reply.  Maybe some of
this can be cleaned up when we start to add more buffer pools.