Update copyright notices.
Where I believe that I wrote the file, remove any year from the file.
This exact year information can easily be determined from the git
version history.  This also means that there is less churn updating the
files from year to year.

Where the copyright declarations are already more complex, just add the
current year to each one.  In the future, the git history should be audited
to confirm there is no discrepancy between the claims and the history,
and if so, the years could be removed, but the names retained.

Either way, the aim is to clearly use the git repository as the source
of knowledge about the contributions for each file and avoid regular
churn to the header comment.