Avoid logspam by refactoring one windows message.
Because, of course, windows has a completely different error reporting
mechanism that is almost - but not quite - entirely different to the
entire rest of the world.  Also, they have an error number that is
almost -  but not quite - entirely different to the standards used by
the rest of the world.

So, work around this with another set of windows specific hacks.