Fix race condition with setting win32 metric
On my test system, a 1 second sleep allowed enough time for the system
to stabilise and set the metric.

My guess is that there is something earlier in the wintap.c that is
using a different interface (perhaps one of the remaining cmd lines?)
and that process is (re)setting the metric after this API call.

This is pretty horrid, but welcome to windows programming.