Note about paginated data.
This has not been implemented yet, but documenting it sooner so that
any consumer can adjust.

Ideally, random empty items would be inserted all the time, but that
would make the implementation simplicity go away.

The current implementation of pagination actually causes the daemon to
iterate over all earlier offsets, which is not efficient.  If the
structures that are used to store the data are changed from the current
linked list of hashed malloc()ed items into an array of hashed items
then a simple process of jump to offset and output limit items can be
implemented.  This array will have some empty slots, so they will still
be returned to the API caller, just as empty dictionaries