Remove unused socket type from n2nsock processing.
This parameter to fill_n2nsock was never being set inside the n2nsock
structure.  The encode and decode routines for the wire representation
of the n2nsock do appear to be correctly handling the value, but when
the fill_n2nsock was updated to set the field correctly, nothing else
worked.

The current TCP handling is mostly done with special-casing the edge, so
when that gets improved, this may come back - but it will probably cause
incompatibilities with n2n due to the observed handling issues.