Make the memory tracking a compile-time option
Having looked closer at the implementation of mallinfo2(), it doesnt
seem like something I would want to run for every metrics export.

Additionally, the feature is only available on glibc, which both makes
it hard to use generically /and/ made it impossible to use to debug the
memory issues being seen in a OpenWRT musl environment.