Actually make musl detectable
In a fit of academic purity, the musl devs are so confident that there
is no reason for anyone to ever need to detect that they are building
for a musl system that they refuse to have a define to test for its use.
    https://wiki.musl-libc.org/faq

In my situation, I recognise that anyone compiling with musl is likely
to be targeting an embedded or cut-down environment.  Highly likely to
be either a constrained resource environment, or the kind of
containerised environment in which you cannot install other tools.

Specifically, I would like to know if the target system is likely to be
able to have gdb installed.

Since the number of users of this software who use OpenWrt and are not
able to ever install gdb is surprisingly large, this turns out to be a
useful datapoint.

I cannot simply test for OpenWrt, since Alpine - and probably other cut
down build environments - also fall into this category.

I wish to be able to output information useful for debugging on those
systems where it is unlikely that a debugger is possible to install.